[bugfix] repairing name of files and renaming duplicated entity
